move all the zmq related files to a single folder so the linkages for ZMQ can be moved to the network library only and not at the top level cmake.
