+++ src/helics/shared_api_library/helicsExport.cpp
@@ -21,7 +21,7 @@ SPDX-License-Identifier: BSD-3-Clause
 #include <mutex>
 #include <vector>
 #ifdef ENABLE_ZMQ_CORE
-#    include "../common/zmqContextManager.h"
+#    include "../network/zmq/zmqContextManager.h"
 #    include "cppzmq/zmq.hpp"
 #endif
